new clang plugin: loopvartoosmall
Idea from bubli - look for loops where the index variable is of such
size that it cannot cover the range revealed by examining the length
part of the condition.

So far, I have only run the plugin up till the VCL module.

Also the plugin deliberately excludes anything more complicated than a
straightforward incrementing for loop.

(cherry picked from commit d5129a9dd68978f9eccdd4597b5b6834557c422a)

+++ l10ntools/source/xrmmerge.cxx
@@ -413,7 +413,7 @@ void XRMResMerge::WorkOnDesc(
         if ( pEntrys ) {
             OString sCur;
             OString sDescFilename = GetAttribute ( rOpenTag, "xlink:href" );
-            for( unsigned int n = 0; n < aLanguages.size(); n++ ){
+            for( size_t n = 0; n < aLanguages.size(); n++ ){
                 sCur = aLanguages[ n ];
                 OString sContent;
                 if ( !sCur.equalsIgnoreAsciiCase("en-US")  &&
@@ -502,7 +502,7 @@ void XRMResMerge::EndOfText(
         MergeEntrys *pEntrys = pMergeDataFile->GetMergeEntrys( pResData );
         if ( pEntrys ) {
             OString sCur;
-            for( unsigned int n = 0; n < aLanguages.size(); n++ ){
+            for( size_t n = 0; n < aLanguages.size(); n++ ){
                 sCur = aLanguages[ n ];
                 OString sContent;
                 if (!sCur.equalsIgnoreAsciiCase("en-US") &&
